refactor(ui): DOMA-8920 moved IconButton to Button.Icon
